requirements.txt
requests
python-dotenv
google-generativeai
gspread
google-auth
flask
